requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
openai==1.54.3
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
pydantic==2.5.0
reportlab==4.0.7
Pillow==10.1.0
requests==2.31.0
numpy>=1.21.0
httpx==0.27.2
